+++ src/activity.js
@@ -1,17 +1,17 @@
 function handleActivityAction(actor, _obj) {
     if (_obj.slug === "scout") {
         const sc = actorFeat(actor, "incredible-scout");
+        const ded = actorFeat(actor, "scout-dedication");
 
         const party = game.actors.filter(a=>a.isOfType("party")).find(a=>a.members.find(b=>b.id===actor.id));
         if (party) {
-            party.members.forEach(tt => {
-                if (!hasEffectBySourceId(tt, sc ? "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6" : "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")) {
-                    setEffectToActor(tt, sc ? "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6" : "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")
+            party.members.filter(a=>a.uuid != actor.uuid).forEach(tt => {
+                if (!hasEffectBySourceId(tt, ded ? 'Compendium.pf2e-action-support.action-support.Item.U7tuKcRePhSu2C2P' : sc ? "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6" : "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")) {
+                    setEffectToActor(tt, ded ? 'Compendium.pf2e-action-support.action-support.Item.U7tuKcRePhSu2C2P' : sc ? "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6" : "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")
                 }
             })
-        } else if (sc && !hasEffectBySourceId(actor, "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6")) {
-            setEffectToActor(actor, "Compendium.pf2e.other-effects.Item.la8rWwUtReElgTS6")
-        } else if (!hasEffectBySourceId(actor, "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")) {
+        }
+        if (!hasEffectBySourceId(actor, "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")) {
             setEffectToActor(actor, "Compendium.pf2e.other-effects.Item.EMqGwUi3VMhCjTlF")
         }
     } else if (_obj.slug === "follow-the-expert") {